# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NFL Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/leaders": {
      "get": {
        "summary": "Get Leaders",
        "operationId": "get_leaders_leaders_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeagueLeader"
                  },
                  "title": "Response Get Leaders Leaders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/wpa-leaders": {
      "get": {
        "summary": "Get Wpa Leaders",
        "operationId": "get_wpa_leaders_wpa_leaders_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WpaLeaders"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/standings": {
      "get": {
        "summary": "Get Standings",
        "operationId": "get_standings_standings_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DivisionStandings"
                  },
                  "title": "Response Get Standings Standings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Search",
        "operationId": "search_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  },
                  "title": "Response Search Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DivisionStandings": {
        "properties": {
          "division": {
            "type": "string",
            "title": "Division"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/StandingsRow"
            },
            "type": "array",
            "title": "Teams"
          }
        },
        "type": "object",
        "required": [
          "division",
          "teams"
        ],
        "title": "DivisionStandings"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeagueLeader": {
        "properties": {
          "player_id": {
            "type": "string",
            "title": "Player Id"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          },
          "headshot_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headshot Url"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played"
          },
          "attempts": {
            "type": "number",
            "title": "Attempts"
          },
          "completions": {
            "type": "number",
            "title": "Completions"
          },
          "pass_yards": {
            "type": "number",
            "title": "Pass Yards"
          },
          "pass_tds": {
            "type": "number",
            "title": "Pass Tds"
          },
          "interceptions_thrown": {
            "type": "number",
            "title": "Interceptions Thrown"
          },
          "sacks_taken": {
            "type": "number",
            "title": "Sacks Taken"
          },
          "pass_epa": {
            "type": "number",
            "title": "Pass Epa"
          },
          "carries": {
            "type": "number",
            "title": "Carries"
          },
          "rush_yards": {
            "type": "number",
            "title": "Rush Yards"
          },
          "rush_tds": {
            "type": "number",
            "title": "Rush Tds"
          },
          "rush_epa": {
            "type": "number",
            "title": "Rush Epa"
          },
          "targets": {
            "type": "number",
            "title": "Targets"
          },
          "receptions": {
            "type": "number",
            "title": "Receptions"
          },
          "rec_yards": {
            "type": "number",
            "title": "Rec Yards"
          },
          "rec_tds": {
            "type": "number",
            "title": "Rec Tds"
          },
          "yac": {
            "type": "number",
            "title": "Yac"
          },
          "air_yards": {
            "type": "number",
            "title": "Air Yards"
          },
          "rec_epa": {
            "type": "number",
            "title": "Rec Epa"
          },
          "solo_tackles": {
            "type": "number",
            "title": "Solo Tackles"
          },
          "assist_tackles": {
            "type": "number",
            "title": "Assist Tackles"
          },
          "tackles_for_loss": {
            "type": "number",
            "title": "Tackles For Loss"
          },
          "sacks": {
            "type": "number",
            "title": "Sacks"
          },
          "qb_hits": {
            "type": "number",
            "title": "Qb Hits"
          },
          "def_interceptions": {
            "type": "number",
            "title": "Def Interceptions"
          },
          "pass_breakups": {
            "type": "number",
            "title": "Pass Breakups"
          },
          "forced_fumbles": {
            "type": "number",
            "title": "Forced Fumbles"
          },
          "fumble_recoveries": {
            "type": "number",
            "title": "Fumble Recoveries"
          },
          "fg_att": {
            "type": "number",
            "title": "Fg Att"
          },
          "fg_made": {
            "type": "number",
            "title": "Fg Made"
          },
          "xp_att": {
            "type": "number",
            "title": "Xp Att"
          },
          "xp_made": {
            "type": "number",
            "title": "Xp Made"
          },
          "punts": {
            "type": "number",
            "title": "Punts"
          },
          "punt_yards": {
            "type": "number",
            "title": "Punt Yards"
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "player_name",
          "position",
          "team",
          "headshot_url",
          "games_played",
          "attempts",
          "completions",
          "pass_yards",
          "pass_tds",
          "interceptions_thrown",
          "sacks_taken",
          "pass_epa",
          "carries",
          "rush_yards",
          "rush_tds",
          "rush_epa",
          "targets",
          "receptions",
          "rec_yards",
          "rec_tds",
          "yac",
          "air_yards",
          "rec_epa",
          "solo_tackles",
          "assist_tackles",
          "tackles_for_loss",
          "sacks",
          "qb_hits",
          "def_interceptions",
          "pass_breakups",
          "forced_fumbles",
          "fumble_recoveries",
          "fg_att",
          "fg_made",
          "xp_att",
          "xp_made",
          "punts",
          "punt_yards"
        ],
        "title": "LeagueLeader",
        "description": "One row per player in /leaders. Stat fields are season totals from\nSUM() over player_game_stats, returned as DOUBLE.\n\nAll stat fields are required-and-non-null: the underlying columns have\nDEFAULT 0, so SUM(...) is always a number (0.0 for non-applicable\nstats like RB pass_yards). Roster-joined fields (position/team) may be\nnull because the join is LEFT and roster data can be missing."
      },
...
      "SearchResult": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "player",
              "team"
            ],
            "title": "Type"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          },
          "headshot_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headshot Url"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id",
          "name",
          "position",
          "team",
          "headshot_url"
        ],
        "title": "SearchResult",
        "description": "A single search hit. `type` discriminates player vs team."
      },
...
      "StandingsRow": {
        "properties": {
          "team": {
            "type": "string",
            "title": "Team"
          },
          "w": {
            "type": "integer",
            "title": "W"
          },
          "l": {
            "type": "integer",
            "title": "L"
          },
          "t": {
            "type": "integer",
            "title": "T"
          },
          "pct": {
            "type": "number",
            "title": "Pct"
          },
          "pf": {
            "type": "integer",
            "title": "Pf"
          },
          "pa": {
            "type": "integer",
            "title": "Pa"
          },
          "home": {
            "type": "string",
            "title": "Home"
          },
          "away": {
            "type": "string",
            "title": "Away"
          },
          "div": {
            "type": "string",
            "title": "Div"
          },
          "strk": {
            "type": "string",
            "title": "Strk"
          },
          "gb": {
            "type": "string",
            "title": "Gb"
          }
        },
        "type": "object",
        "required": [
          "team",
          "w",
          "l",
          "t",
          "pct",
          "pf",
          "pa",
          "home",
          "away",
          "div",
          "strk",
          "gb"
        ],
        "title": "StandingsRow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WpaLeader": {
        "properties": {
          "player_id": {
            "type": "string",
            "title": "Player Id"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          },
          "headshot_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headshot Url"
          },
          "wpa": {
            "type": "number",
            "title": "Wpa"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played"
          },
          "attempts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempts"
          },
          "carries": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carries"
          },
          "receptions": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receptions"
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "player_name",
          "position",
          "team",
          "headshot_url",
          "wpa",
          "games_played"
        ],
        "title": "WpaLeader",
        "description": "One row in passing/rushing/receiving WPA leaders.\n\nThe volume column (attempts/carries/receptions) varies per sub-list.\nOnly one is populated per row \u2014 the others are simply absent from\nthe SQL projection, so they remain optional + None."
      },
      "WpaLeaders": {
        "properties": {
          "passing": {
            "items": {
              "$ref": "#/components/schemas/WpaLeader"
            },
            "type": "array",
            "title": "Passing"
          },
          "rushing": {
            "items": {
              "$ref": "#/components/schemas/WpaLeader"
            },
            "type": "array",
            "title": "Rushing"
          },
          "receiving": {
            "items": {
              "$ref": "#/components/schemas/WpaLeader"
            },
            "type": "array",
            "title": "Receiving"
          }
        },
        "type": "object",
        "required": [
          "passing",
          "rushing",
          "receiving"
        ],
        "title": "WpaLeaders",
        "description": "The full /wpa-leaders payload: three sub-lists."
      }
    }
  }
}